google_trends/requirements.txt
pytrends
pandas
google-cloud-firestore